splitting the header in specific headers for source and sink only necessary defs/refs for both are now in journal-gateway-zmtp.h
