+++ misc/journal-gateway-zmtp-sink.conf
@@ -1,2 +1,2 @@
 JOURNAL_REMOTE_DIR="/var/log/journal/remote/"
-EXPOSED_SOCKET="tcp://*:5555"
+GATEWAY_LOG_PEER="tcp://*:5555"
